requirements.txt
pandas>=2.0
numpy>=1.24
matplotlib>=3.7
seaborn>=0.12
scikit-learn>=1.3
lightgbm>=4.0
interpret>=0.6
torch>=2.0
psycopg2-binary>=2.9
jupyter>=1.0
